requirements.txt
asgiref==3.5.2
astroid==2.11.5
certifi==2022.5.18.1
colorama==0.4.4
dill==0.3.5.1
distlib==0.3.4
dj-database-url==1.0.0
Django==4.0.4
django-ckeditor==6.5.0
django-heroku==0.3.1
django-js-asset==2.0.0
filelock==3.7.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
Pillow==9.2.0
pipenv==2022.5.2
platformdirs==2.5.2
psycopg2==2.9.3
pylint==2.13.9
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.1
virtualenv==20.14.1
virtualenv-clone==0.5.7
whitenoise==6.2.0
wrapt==1.14.1
